test: add tests for WarningList
