+++ app/tests/components/warningList.nuxt.spec.ts
@@ -1,7 +1,7 @@
 import { expect, it } from "vitest";
 import { mountSuspended, mockNuxtImport } from "@nuxt/test-utils/runtime";
 import { createTask } from "./createTask";
-import WarningsSlideover from "~/components/WarningsSlideover.vue";
+import WarningList from "~/components/WarningList.vue";
 import { TokioTask } from "~/types/task/tokioTask";
 import { Timestamp } from "~/types/common/duration";
 import { NeverYielded } from "~/types/warning/taskWarnings/neverYielded";
@@ -14,14 +14,11 @@ mockNuxtImport("useTasks", () => {
         task.warnings.push(new NeverYielded());
         tasksData.value.set(1n, task);
         const lastUpdatedAt = ref<Timestamp>(new Timestamp(1000n, 0));
-
         return { pending, tasksData, lastUpdatedAt };
     };
 });
 
 it("WarningsSlideover View", async () => {
-    const component = await mountSuspended(WarningsSlideover);
-    const slideover = useSlideover();
-    slideover.open(WarningsSlideover);
+    const component = await mountSuspended(WarningList);
     expect(component.html()).toMatchSnapshot();
 });
